Improved File Structure
