+++ README.md
@@ -1,13 +1,14 @@
 # FischisDiscordBot
-A small German Discord Bot using [Node](https://nodejs.org/) and the [Discord.js](https://discord.js.org) library.
+A small Discord Bot using [Node](https://nodejs.org/) and the [Discord.js](https://discord.js.org) library.
 
-#### WARNING: This Project is my first Node project and also my first Discord Bot, so expect Spaghetti Code and more! 
+#### WARNING: This Project is my first Node project and also my first Discord Bot, so expect Spaghetti Code and more!
+
+##### ALSO WARNING: This bot is German, but we plan to add English in the far Future.
 
 Its current features are:
 1. Getting Hot Subreddit Posts
 2. Basic Fun commands
 
 This Bot is Currently hosted on Glitch.com, so I included all the Glitch Files. To make sure that your Bot does'nt shut down itself. you 
-will need a Website like [Uptime Robot](https://uptimerobot.com/) if you want to learn more about it you can use  [this]
-(https://anidiotsguide_old.gitbooks.io/discord-js-bot-guide/content/other-guides/hosting-on-glitchcom.html) awesome Tutorial, which 
+will need a Website like [Uptime Robot](https://uptimerobot.com/). If you want to learn more about it you can use [this](https://anidiotsguide_old.gitbooks.io/discord-js-bot-guide/content/other-guides/hosting-on-glitchcom.html) awesome Tutorial, which 
 Inspired me to do this thing!
