+++ package.json
@@ -3,7 +3,7 @@
   "//2": "https://docs.npmjs.com/files/package.json",
   "//3": "updating this file will download and update your packages",
   "name": "hello-express",
-  "version": "0.0.1",
+  "version": "0.2.1",
   "description": "A simple Node app built on Express, instantly up and running.",
   "main": "index.js",
   "scripts": {
